UI improvements: - current section is displayed as active - change logo link with canister id for firefox - identity saved in page and stored in localstorage - create license with Internet Identity login - fix loader, license name filter - init paypal only when needed - logout

+++ src/licenseManager/lib.rs
@@ -607,7 +607,7 @@ struct License {
     pub perpetual: bool,
     pub duration: u64,
     pub transferable: bool,
-    pub transfer_commission: u64
+    pub transfer_commission: f64
     // pub keywords: Vec<String>,
 }
